Add documentation for ECMA-167 / ECMA-168; Begin adding required structures

+++ TotalImage.IO/FileSystems/ISO/IsoBootVolumeDescriptor.cs
@@ -32,7 +32,7 @@ namespace TotalImage.FileSystems.ISO
         /// <param name="identifier">The volume descriptor identifier</param>
         /// <param name="version">The version of the volume descriptor</param>
         public IsoBootVolumeDescriptor(in ReadOnlySpan<byte> record, in IsoVolumeDescriptorType type, in ImmutableArray<byte> identifier, in byte version)
-            : base(type, identifier, version)
+            : base((byte)type, identifier, version)
         {
             Span<char> tmp = new char[32];
