+++ TotalImage.IO/FileSystems/ISO/IsoUnknownVolumeDescriptor.cs
@@ -22,7 +22,7 @@ namespace TotalImage.FileSystems.ISO
         /// <param name="identifier">The volume descriptor identifier</param>
         /// <param name="version">The version of the volume descriptor</param>
         public IsoUnknownVolumeDescriptor(in ReadOnlySpan<byte> record, in IsoVolumeDescriptorType type, in ImmutableArray<byte> identifier, in byte version)
-            : base(type, identifier, version)
+            : base((byte)type, identifier, version)
         {
             if (identifier.SequenceEqual(HsfStandardIdentifier))
                 Content = record[15..].ToArray().ToImmutableArray();
